+++ src/TitleDetails/AddPieceModal/AddPieceModal.js
@@ -265,7 +265,7 @@ const AddPieceModal = ({
           <Row>
             <Col xs={6}>
               <FieldInventory
-                instanceId={instanceId}
+                instanceId={isLocationRequired ? instanceId : undefined}
                 locationIds={locationIds}
                 locations={locations}
 
@@ -349,8 +349,9 @@ export default stripesFinalForm({
   mutators: {
     setLocationValue: (args, state, tools) => {
       const [location, locationField, holdingFieldName, holdingId] = args;
+      const locationId = holdingId ? undefined : location?.id || location;
 
-      tools.changeValue(state, locationField, () => location?.id || location);
+      tools.changeValue(state, locationField, () => locationId);
 
       if (holdingFieldName) {
         tools.changeValue(state, holdingFieldName, () => holdingId);