+++ src/TitleReceive/TitleReceive.js
@@ -131,8 +131,9 @@ export default stripesFinalForm({
   mutators: {
     setLocationValue: (args, state, tools) => {
       const [location, locationField, holdingFieldName, holdingId] = args;
+      const locationId = holdingId ? undefined : location?.id || location;
 
-      tools.changeValue(state, locationField, () => location?.id || location);
+      tools.changeValue(state, locationField, () => locationId);
 
       if (holdingFieldName) {
         tools.changeValue(state, holdingFieldName, () => holdingId);