+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Display Caption, Copy number in piece form. Refs UIREC-160.
 * Update Received/Expected piece table columns. Refs UIREC-158.
 * Receiving - Implement MCL Next/Previous pagination. Refs UIREC-164.
+* Piece location/holding field adjustments. Refs UIREC-162.
 * Display Holding OR Location name in location column in search results. Refs UIREC-172.
 * Filter label contains extra 's'. Refs UIREC-173.
 * Disable add and remove piece when order is pending. Refs UIREC-174.